refactor: handle_animation to util
